change precision of frame timer. Hopefully that helps the rounding issues.

+++ include/Cala/backend/vulkan/Driver.h
@@ -57,9 +57,9 @@ namespace cala::backend::vulkan {
 
         u32 setLayoutCount() const { return _setLayouts.size(); }
 
-        f32 fps() const { return 1000.f / (static_cast<f32>(_lastFrameTime.microseconds()) / 1000.f); }
+        f64 fps() const { return 1000.f / (static_cast<f64>(_lastFrameTime.microseconds()) / 1000.f); }
 
-        f32 milliseconds() const { return static_cast<f32>(_lastFrameTime.microseconds()) / 1000.f; }
+        f64 milliseconds() const { return static_cast<f64>(_lastFrameTime.microseconds()) / 1000.f; }
 
     private:
